+++ MatrixSDK/Utils/MXBackgroundTask.h
@@ -1,14 +1,12 @@
 /*
- Copyright 2017 Samuel Gallet
- Copyright 2017 Vector Creations Ltd
  Copyright 2019 The Matrix.org Foundation C.I.C
-
+ 
  Licensed under the Apache License, Version 2.0 (the "License");
  you may not use this file except in compliance with the License.
  You may obtain a copy of the License at
-
+ 
  http://www.apache.org/licenses/LICENSE-2.0
-
+ 
  Unless required by applicable law or agreed to in writing, software
  distributed under the License is distributed on an "AS IS" BASIS,
  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -16,25 +14,33 @@
  limitations under the License.
  */
 
-#ifndef MXBackgroundModeHandler_h
-#define MXBackgroundModeHandler_h
+#ifndef MXBackgroundTask_h
+#define MXBackgroundTask_h
 
 #import <Foundation/Foundation.h>
-#import "MXBackgroundTask.h"
-
-typedef void (^MXBackgroundTaskExpirationHandler)(void);
 
 NS_ASSUME_NONNULL_BEGIN
 
-/**
- Interface to handle enabling background mode
- */
-@protocol MXBackgroundModeHandler <NSObject>
+typedef void (^MXBackgroundTaskExpirationHandler)(void);
 
-- (id<MXBackgroundTask>)startBackgroundTaskWithName:(NSString *)name expirationHandler:(nullable MXBackgroundTaskExpirationHandler)expirationHandler;
+/**
+ MXBackgroundTask is protocol describing a background task regardless of the platform used.
+ */
+@protocol MXBackgroundTask <NSObject>
+
+// Name of the background task for debug.
+@property (nonatomic, strong, readonly) NSString *name;
+
+// Yes if the background task is currently running.
+@property (nonatomic, readonly) BOOL isRunning;
+
+/**
+ Stop the background task. Cannot be started anymore.
+ */
+- (void)stop;
 
 @end
 
 NS_ASSUME_NONNULL_END
 
-#endif /* MXBackgroundModeHandler_h */
+#endif /* MXBackgroundTask_h */