+++ CHANGES.rst
@@ -5,6 +5,7 @@ Improvements:
  * MX3PidAddManager: Add User-Interactive Auth to /account/3pid/add (vector-im/riot-ios#2744).
  * MXSession: On resume, make the first /sync request trigger earlier (vector-im/riot-ios#2793).
  * MXCrypto: Do not fail to decrypt when there is nothing to decrypt (redacted events).
+ * MXBackgroundModeHandler: Update interface and now return a MXBackgroundTask, a protocol describing a background task regardless of the plartform used.
 
 Bug fix:
  * Room members who left are listed with the actual members (vector-im/riot-ios#2737).
@@ -12,6 +13,9 @@ Bug fix:
  * MXHTTPOperation: Make urlResponseFromError return the url response in case of MXError.
  * MXHTTPOperation: Fix a crash in `-mutateTo:` method when operation parameter is nil.
  * VoIP: Fix regression when using a TURN server (vector-im/riot-ios#2796).
+
+API break:
+ * MXBackgroundModeHandler: Update interface and now use a single method that return a MXBackgroundTask.
 
 Changes in Matrix iOS SDK in 0.14.0 (2019-10-11)
 ===============================================